second commit

+++ HTTTQL_NhanSu_Client/src/main/java/qlns/Luong.java
@@ -19,5 +19,5 @@ public class Luong {
 	
 	private String tenLuong;
 	
-	private float mucLuong;
+	private int mucLuong;
 }
